[REFACT] 커스텀 hook 생성하기, React.memo로 성능 최적화하기 #NADA-50
